refactor(ui): extract herd/HerdEpicGroups + HerdDoneList + HerdEmptyState from Herd (#855)
