+++ ui/src/lib/components/Herd.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import type { Session, GitState, SessionActivity, Epic, CompletedEpic } from "$lib/types";
   import type { BlockState } from "$lib/triage";
-  import UnitRow from "./UnitRow.svelte";
   import HerdGroup from "./herd/HerdGroup.svelte";
   import type { HerdRowCtx } from "./herd/HerdGroup.svelte";
   import HerdFilterBar from "./herd/HerdFilterBar.svelte";
   import HerdSegRow from "./herd/HerdSegRow.svelte";
-  import EmptyHerd from "./EmptyHerd.svelte";
-  import EpicGroupHeader from "./EpicGroupHeader.svelte";
+  import HerdEpicGroups from "./herd/HerdEpicGroups.svelte";
+  import HerdDoneList from "./herd/HerdDoneList.svelte";
+  import HerdEmptyState from "./herd/HerdEmptyState.svelte";
   import IntegratedEpicsBand from "./IntegratedEpicsBand.svelte";
   import RundownPanel from "./RundownPanel.svelte";
   import { partitionSessions, shownSessions, type HerdFilter } from "./herd-partition";
@@ -15,9 +15,6 @@
   import { collectReadyPrs } from "./merge-train";
   import { displayStatus } from "$lib/display-status";
   import { reviews, planGates } from "$lib/reviews.svelte";
-  import { recaps } from "$lib/recaps.svelte";
-  import { formatAgo } from "$lib/format";
-  import type { RecapVerdict } from "$lib/types";
   import { m } from "$lib/paraglide/messages";
 
   let {
@@ -394,24 +391,6 @@
       },
     ].filter(Boolean) as PartitionGroupEntry[],
   );
-
-  // Done lens row chrome: a finished session's recap verdict drives a small chip. Semantic
-  // colors mirror SessionRecap/DoneRecapPanel (green = genuinely READY only; parked = slate;
-  // needs-attention = amber).
-  const VERDICT_COLOR: Record<RecapVerdict, string> = {
-    ready: "var(--color-green)",
-    parked: "var(--status-done)",
-    needs_attention: "var(--color-amber)",
-  };
-  function verdictLabel(v: RecapVerdict): string {
-    if (v === "ready") return m.recap_verdict_ready();
-    if (v === "parked") return m.recap_verdict_parked();
-    return m.recap_verdict_needs_attention();
-  }
-  // last path segment of a repoPath, for the done row's repo label
-  function repoBasename(p: string): string {
-    return p.split("/").filter(Boolean).at(-1) ?? p;
-  }
 </script>
 
 <div class="panel bracket" class:flow>
@@ -435,96 +414,42 @@
     {:else if filter === "done"}
       <!-- Done lens: archived sessions from the page's lazy doneSessions store (NOT the
          live `sessions` list). Read-only rows; clicking opens the DoneRecapPanel. -->
-      {#if doneList.length === 0}
-        <div class="empty micro static">{m.herd_done_empty()}</div>
-      {:else}
-        {#each doneList as ds (ds.id)}
-          {@const r = recaps.map[ds.id]}
-          <button
-            type="button"
-            class="done-row"
-            class:sel={ds.id === doneSelectedId}
-            data-unit-id={ds.id}
-            onclick={() => ondoneselect?.(ds.id)}
-          >
-            <div class="done-row-top">
-              <span class="done-desig">{ds.desig}</span>
-              <span class="done-repo" title={ds.repoPath}>{repoBasename(ds.repoPath)}</span>
-              {#if r?.state === "ready" && r.verdict}
-                <span class="done-verdict" style:color={VERDICT_COLOR[r.verdict]}
-                  >{verdictLabel(r.verdict)}</span
-                >
-              {/if}
-              <span class="done-ago"
-                >{m.done_recap_finished({
-                  ago: formatAgo(nowMs - (ds.archivedAt ?? ds.updatedAt)),
-                })}</span
-              >
-            </div>
-            <span class="done-snippet"
-              >{r?.state === "ready" ? r.headline : ds.name || ds.prompt}</span
-            >
-          </button>
-        {/each}
-      {/if}
+      <HerdDoneList {doneList} {doneSelectedId} {ondoneselect} {nowMs} />
     {:else if sessions.length === 0}
       <!-- the status filter empties the list at PAGE level (herdSessions), so an
          empty status result lands HERE — it must outrank the repo note and the
          first-run EmptyHerd nudge, and name the active status so vanished
          done/parked sessions read as intentional -->
-      {#if statusFilter != null && filteredRepo}
-        <div class="empty micro static">
-          {m.herd_status_repo_filter_empty({ status: statusLabel, repo: filteredRepo })}
-        </div>
-      {:else if statusFilter != null}
-        <div class="empty micro static">
-          {m.herd_status_filter_empty({ status: statusLabel })}
-        </div>
-      {:else if filteredRepo}
-        <div class="empty micro static">{m.herd_repo_filter_empty({ repo: filteredRepo })}</div>
-      {:else}
-        <EmptyHerd {onnew} {issueActionsUnset} {onsettings} />
-      {/if}
+      <HerdEmptyState
+        mode="sessions"
+        {statusFilter}
+        {statusLabel}
+        {filteredRepo}
+        {filter}
+        {issueActionsUnset}
+        {onnew}
+        {onsettings}
+      />
     {:else if shown.length === 0}
-      {#if filter === "research"}
-        <div class="empty micro static">{m.herd_research_empty()}</div>
-      {:else}
-        <div class="empty micro static">{m.herd_ready_empty()}</div>
-      {/if}
+      <HerdEmptyState
+        mode="shown"
+        {statusFilter}
+        {statusLabel}
+        {filteredRepo}
+        {filter}
+        {issueActionsUnset}
+        {onnew}
+        {onsettings}
+      />
     {:else}
-      {#each grouped.groups as g (g.key)}
-        <EpicGroupHeader
-          epic={g.epic}
-          collapsed={collapsedKeys.has(g.key)}
-          cues={cuesFor(g)}
-          ontoggle={() => oncollapsetoggle?.(g.key)}
-          {onepic}
-        />
-        {#if !collapsedKeys.has(g.key)}
-          <div class="epic-children">
-            {#each g.sessions as session (session.id)}
-              <UnitRow
-                {session}
-                selected={session.id === selectedId}
-                {nowMs}
-                {onselect}
-                git={git[session.id]}
-                activity={activity[session.id]}
-                previewPort={preview[session.id] ?? null}
-                previewServeFailed={previewServe[session.id] === "failed"}
-                {onpreview}
-                {ondecommission}
-                {onrelaunch}
-                {onrelaunchElsewhere}
-                {repoFilter}
-                {onrepofilter}
-                {workingBlocked}
-                quotaKind={quotaKindFor(session.id)}
-              />
-            {/each}
-          </div>
-        {/if}
-      {/each}
+      <HerdEpicGroups
+        groups={grouped.groups}
+        {collapsedKeys}
+        {cuesFor}
+        {onepic}
+        {oncollapsetoggle}
+        ctx={rowCtx}
+      />
       {#each partitionGroups as grp (grp.key)}
         <HerdGroup ctx={rowCtx} {...grp} />
       {/each}
@@ -610,15 +535,6 @@
     color: var(--color-muted);
   }
 
-  /* Child rows of an epic group sit lightly inset under their headline so the
-     group reads as one unit. A hairline rail on the leading edge reinforces the
-     nesting without a heavy indent. Token-based; no raw px color. */
-  .epic-children {
-    padding-left: 10px;
-    margin-left: 4px;
-    border-left: 1px solid color-mix(in srgb, var(--color-blue) 30%, var(--color-line));
-  }
-
   .units {
     overflow: auto;
     padding: 6px;
@@ -648,93 +564,5 @@
   }
   .units.flow :global(.unit.sel::after) {
     display: none;
-  }
-
-  .empty {
-    width: 100%;
-    padding: 24px 14px;
-    text-align: center;
-    color: var(--color-faint);
-    border: 0;
-    background: none;
-    font-family: inherit;
-    cursor: pointer;
-    transition: color 0.12s ease;
-  }
-  .empty:hover {
-    color: var(--color-ink);
-  }
-  .empty.static {
-    cursor: default;
-  }
-  .empty.static:hover {
-    color: var(--color-faint);
-  }
-
-  /* Done-lens row: a finished session, picked to show its recap in the main panel.
-     Borrows the rail's row rhythm/tokens; selection mirrors UnitRow's .sel cue. */
-  .done-row {
-    display: flex;
-    flex-direction: column;
-    gap: 3px;
-    width: 100%;
-    text-align: left;
-    border: 1px solid transparent;
-    background: none;
-    font-family: inherit;
-    cursor: pointer;
-    padding: 8px 10px;
-    transition:
-      border-color 0.12s ease,
-      background 0.12s ease;
-  }
-  .done-row:hover {
-    border-color: var(--color-line);
-    background: var(--color-hover);
-  }
-  .done-row:focus-visible {
-    outline: none;
-    box-shadow: inset 0 0 0 1px var(--color-amber);
-  }
-  .done-row.sel {
-    border-color: var(--color-line-bright);
-    background: var(--color-sel);
-  }
-  .done-row-top {
-    display: flex;
-    align-items: baseline;
-    gap: 8px;
-  }
-  .done-desig {
-    font-size: var(--fs-meta);
-    color: var(--color-ink-bright);
-    font-weight: 600;
-  }
-  .done-repo {
-    font-size: var(--fs-micro);
-    color: var(--color-muted);
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-  .done-verdict {
-    font-size: var(--fs-micro);
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.04em;
-    flex-shrink: 0;
-  }
-  .done-ago {
-    margin-left: auto;
-    flex-shrink: 0;
-    font-size: var(--fs-micro);
-    color: var(--color-faint);
-  }
-  .done-snippet {
-    font-size: var(--fs-meta);
-    color: var(--color-ink);
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
   }
 </style>
